Updated to 2026
Charts now show 2026-2031, free agents after 2025 are still visible on previous team, dead cap hits that ended after 2025 are gone. Should find a way to streamline this change for the future - had to change things in a handful of different places.

Also changed injury tags to show if a player is listed as OUT (red flag with line through it) or if he's on the injury list, since those designations won't happen until closer to opening day.